Use configurations instead of cross-referencing project tasks
Gradle recommends [1] that artifacts be shared across projects by using
configurations: a consumable configuration on the producer side, and a
resolvable configuration on the consumer side.

This is preferable as it allows Gradle to control and handle the
artifact sharing safely, instead of the unsafe way of directly reaching
across projects and accessing other projects' tasks.

[1]: https://docs.gradle.org/current/userguide/cross_project_publications.html